modified controller inputs
I changed the inputs for the up/down and rotate functions.  Up/Down is now up/down on the right analog stick of the ps4 controller. Rotate is now L/R on the right analog stick.  The operation becomes a little more intuitive, but it can be challenging for precision. Recommend trying to adjust a larger deadzone for the right analog stick...